🔨 refactor: 테스트 코드 디렉토리 구조 분리
- 기존 test/domain/{domain name} 구조에서 테스트 단위별로 분리
- test/unit/domain/{domain name}
- test/intergration/domain/{domain name}
